+++ src/main/java/kr/codesquad/Main.java
@@ -3,7 +3,7 @@ package kr.codesquad;
 public class Main {
 
   public static void main(String[] args) {
-    new App().run();
+    new ConsoleApp().run();
   }
 
 }
