+++ src/main/java/kr/codesquad/common/exception/ExceptionHandler.java
@@ -1,0 +1,13 @@
+package kr.codesquad.common.exception;
+
+public class ExceptionHandler {
+
+  public void runWithHandler(Runnable runnable) {
+    try {
+      runnable.run();
+    } catch (NumberFormatException e) {
+
+    }
+  }
+
+}
